--- a/cmd/xl-v1-healing-common.go
+++ b/cmd/xl-v1-healing-common.go
@@ -171,20 +171,15 @@ func disksWithAllParts(ctx context.Context, onlineDisks []StorageAPI, partsMetad
 
 		switch scanMode {
 		case madmin.HealDeepScan:
-			erasureInfo := partsMetadata[i].Erasure
-			erasure, err := NewErasure(ctx, erasureInfo.DataBlocks, erasureInfo.ParityBlocks, erasureInfo.BlockSize)
-			if err != nil {
-				dataErrs[i] = err
-				continue
-			}
+			erasure := partsMetadata[i].Erasure
 
 			// disk has a valid xl.json but may not have all the
 			// parts. This is considered an outdated disk, since
 			// it needs healing too.
 			for _, part := range partsMetadata[i].Parts {
-				checksumInfo := erasureInfo.GetChecksumInfo(part.Number)
+				checksumInfo := erasure.GetChecksumInfo(part.Number)
 				partPath := pathJoin(object, fmt.Sprintf("part.%d", part.Number))
-				err = onlineDisk.VerifyFile(bucket, partPath, erasure.ShardFileSize(part.Size), checksumInfo.Algorithm, checksumInfo.Hash, erasure.ShardSize())
+				err := onlineDisk.VerifyFile(bucket, partPath, erasure.ShardFileSize(part.Size), checksumInfo.Algorithm, checksumInfo.Hash, erasure.ShardSize())
 				if err != nil {
 					if !IsErr(err, []error{
 						errFileNotFound,