## Description

* Update reedsolomon package
* Only create encoder when strictly needed.

Warp:
```
λ warp cmp before.csv.zst after.csv.zst
-------------------
Operation: PUT
Operations: 302 -> 311
Duration: 46s -> 48s
* Average: +4.22% (+18.3 MiB/s) throughput, +4.22% (+0.2) obj/s
* Fastest: +6.20% (+45.4 MiB/s) throughput, +6.20% (+0.5) obj/s
* 50% Median: +7.71% (+30.6 MiB/s) throughput, +7.71% (+0.3) obj/s
* Slowest: -3.93% (-13.7 MiB/s) throughput, -3.93% (-0.1) obj/s
```

This is heavily IO  limited on my system, so difference is likely random variation. Especially since RS is only about 0.12% of CPU.

However, comparing pprof of the same run, even with increased throughput, CPU usage was less:

Before:

```
Active filters:
   focus=reedsolomon
Showing nodes accounting for 7.98s, 0.12% of 6617.56s total
      flat  flat%   sum%        cum   cum%
[...]
     0.12s 0.0018%  0.12%      7.87s  0.12%  github.com/klauspost/reedsolomon.reedSolomon.codeSomeShardsP.func1

```

After:

```
Active filters:
   focus=reedsolomon
Showing nodes accounting for 7.64s, 0.11% of 6673.83s total
      flat  flat%   sum%        cum   cum%
[...]
     0.05s 0.00075%  0.11%      7.53s  0.11%  github.com/klauspost/reedsolomon.reedSolomon.codeSomeShardsP.func1

```

7.98s -> 7.64s with more throughput.


## How to test this PR?

Run the server.

## Types of changes
- [x] Optimization