@minio-trusted 2020-05-09
## Mint Automation
| Test | Result |
|:--- |:--- |
| mint-xl.sh | :heavy_check_mark: |
| mint-large-bucket.sh | :heavy_check_mark: |
| mint-fs.sh | :heavy_check_mark: |
| mint-dist-xl.sh | :heavy_check_mark: |
| mint-gateway-nas.sh | :heavy_check_mark: |
| mint-gateway-s3.sh | :x: [more...](#07e7685c-8337-4651-b24a-5211d6618d95) |
| mint-gateway-azure.sh | :x: [more...](#f5fd3a98-a5e0-45fe-8477-92e18cfd573e) |
<a name="07e7685c-8337-4651-b24a-5211d6618d95"></a>
### 9562-8236cd3/mint-gateway-s3.sh.log:
```
Running with
SERVER_ENDPOINT:      minio-dev4.minio.io:32358
ACCESS_KEY:           minio
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 3c3d0b4d7da7:/mint/log /tmp/mint-logs'

(1/15) Running aws-sdk-go tests ... done in 0 seconds
(2/15) Running aws-sdk-java tests ... done in 1 seconds
(3/15) Running aws-sdk-php tests ... done in 42 seconds
(4/15) Running aws-sdk-ruby tests ... done in 1 seconds
(5/15) Running awscli tests ... FAILED in 20 seconds
{
  "name": "awscli",
  "duration": 4364,
  "function": "aws --endpoint-url http://minio-dev4.minio.io:32358 s3api get-object --bucket awscli-mint-test-bucket-30395 --key awscli-mint-test-bucket-30395-object /tmp/datafile-0-b\n",
  "status": "FAIL",
  "error": "Etag mismatch for multipart 0 byte object"
}
(5/15) Running healthcheck tests ... done in 0 seconds
(6/15) Running mc tests ... done in 18 seconds
(7/15) Running minio-dotnet tests ... done in 28 seconds
(8/15) Running minio-go tests ... done in 25 seconds
(9/15) Running minio-java tests ... done in 30 seconds
(10/15) Running minio-js tests ... done in 36 seconds
(11/15) Running minio-py tests ... done in 1 minutes and 15 seconds
(12/15) Running s3cmd tests ... done in 1 minutes and 4 seconds
(13/15) Running s3select tests ... done in 1 seconds
(14/15) Running security tests ... done in 0 seconds

Executed 14 out of 15 tests successfully.
```
<a name="f5fd3a98-a5e0-45fe-8477-92e18cfd573e"></a>
### 9562-8236cd3/mint-gateway-azure.sh.log:
```
Running with
SERVER_ENDPOINT:      minio-dev7.minio.io:30005
ACCESS_KEY:           minioazure
SECRET_KEY:           ***REDACTED***
ENABLE_HTTPS:         0
SERVER_REGION:        us-east-1
MINT_DATA_DIR:        /mint/data
MINT_MODE:            full
ENABLE_VIRTUAL_STYLE: 0

To get logs, run 'docker cp 303f236dc9fa:/mint/log /tmp/mint-logs'

(1/15) Running aws-sdk-go tests ... done in 8 seconds
(2/15) Running aws-sdk-java tests ... done in 2 seconds
(3/15) Running aws-sdk-php tests ... done in 4 minutes and 46 seconds
(4/15) Running aws-sdk-ruby tests ... done in 16 seconds
(5/15) Running awscli tests ... done in 2 minutes and 41 seconds
(6/15) Running healthcheck tests ... done in 0 seconds
(7/15) Running mc tests ... done in 3 minutes and 45 seconds
(8/15) Running minio-dotnet tests ... done in 2 minutes and 19 seconds
(9/15) Running minio-go tests ... done in 3 minutes and 53 seconds
(10/15) Running minio-java tests ... done in 25 minutes and 52 seconds
(11/15) Running minio-js tests ... done in 2 minutes and 57 seconds
(12/15) Running minio-py tests ... done in 9 minutes and 31 seconds
(13/15) Running s3cmd tests ... FAILED in 42 seconds
{
  "name": "s3cmd",
  "duration": "1742",
  "function": "test_copy_object",
  "status": "FAIL",
  "error": ""
}
(13/15) Running s3select tests ... done in 21 seconds
(14/15) Running security tests ... done in 0 seconds

Executed 14 out of 15 tests successfully.
```
Deleting image on docker hub
Deleting image locally